[MOBILE] [FIX] Fix the styles for menu pages so it's complient with the rest of the mobile architecture
