+++ packages/restoMobile/src/pages/MenuPage/MenuPage.styles.ts
@@ -1,0 +1,64 @@
+import { StyleSheet, Dimensions, Platform } from "react-native";
+
+const deviceWidth = Math.round(Dimensions.get('window').width);
+const offset = 40;
+const radius = 20;
+const iconSize = 20;
+
+const styles = StyleSheet.create({
+  
+  container: {
+    width: deviceWidth - 20,
+    alignItems: 'center',
+    marginTop: 25,
+    marginBottom: 25,
+  },
+  scrollView: {
+    alignItems: 'center',
+  },
+  card: {
+    width: deviceWidth - offset,
+    backgroundColor: '#FFFFFF',
+    height: 230,
+    margin: 20,
+    borderRadius: radius,
+    shadowColor: '#000',
+    shadowOffset: {
+      width: 2,
+      height: 2,
+    },
+    shadowOpacity: 0.75,
+    shadowRadius: 4,
+    elevation: 9,
+    position: 'relative',
+    overflow: 'hidden',
+  },
+  cardImage: {
+    height: 130,
+    width: deviceWidth - offset,
+    borderTopLeftRadius: radius,
+    borderTopRightRadius: radius,
+    opacity: 0.9,
+    alignContent: 'center',
+    alignSelf: 'center',
+  },
+  cardContent: {
+    padding: 10,
+  },
+  cardTitle: {
+    fontSize: 18,
+    fontWeight: 'bold',
+  },
+  iconContainer: {
+    flexDirection: 'row',
+    justifyContent: 'space-between',
+    padding: 10,
+  },
+  groupTitle: {
+    fontSize: 20,
+    fontWeight: 'bold',
+    marginVertical: 10,
+  },
+});
+
+export default styles;
